requirements.txt
mock==1.3.0
pytest==2.8.5
python-decouple==3.0
tox==2.3.1
